Split it from ddl
