+++ pkg/statistics/handle/ddl/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "ddl",
-    srcs = ["ddl.go"],
+    srcs = [
+        "ddl.go",
+        "exchange_partition.go",
+    ],
     importpath = "github.com/pingcap/tidb/pkg/statistics/handle/ddl",
     visibility = ["//visibility:public"],
     deps = [
